+++ api/sms/utils.py
@@ -2,7 +2,7 @@ from vms.models import DefaultDc
 # SMS services
 # Must be imported this way and the views modules should have a sms_send function and PROVIDER_NAME constant
 import api.sms.smsapi.views
-
+from api.sms.exceptions import InvalidSMSService
 
 SMS_SERVICES = (
     ('smsapi', api.sms.smsapi.views),
@@ -24,4 +24,7 @@ def get_current_service(settings=None):
     if not settings:
         settings = DefaultDc().settings  # dc1_settings
 
-    return dict(SMS_SERVICES)[settings.SMS_PREFERRED_SERVICE]
+    try:
+        return dict(SMS_SERVICES)[settings.SMS_PREFERRED_SERVICE]
+    except KeyError:
+        raise InvalidSMSService('Preferred SMS service "%s" not found' % settings.SMS_PREFERRED_SERVICE)